프로필 수정을 제외한 유저 기능 완료 (#1)
* :tada: init related user

* :recycle: UserDetailsImpl안에 있는 User를 Dto로 변환

* :recycle: Position Entity와 연관관계 설정

* :recycle: id Long타입에서 Short로 변경

* :recycle: UserId -> userId

* :recycle: memberId -> userId

* :recycle: positionId 타입 Long에서 Short로 변경

* :recycle: memberId -> userId

* :sparkles: ErrorCode 생성

* :recycle: 인자값 변경

* :recycle: 인자값 변경

* :recycle: Long -> Short로 변경

* :recycle: 필수 어노테이션 추가

* :sparkles: 패스워드 유효성 검사 어노테이션 생성

* :sparkles: 패스워드 유효성 검사 어노테이션 생성

* :recycle: 기존 패스워드 유효성 어노테이션 -> 커스텀 어노테이션으로 변경

* :sparkles: 패스워드 변경, 프로필 조회, 회원탈퇴 기능 구현

* :wrench: PASSWORD_NOT_MATCH 메시지 추가

* :recycle: ExceptionHandler 추가

+++ src/main/java/org/example/develociraptor/global/idclass/PositionProjectId.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Getter
 public class PositionProjectId {
-	private Long positionId;
+	private Short positionId;
 	private Long projectId;
 }
